+++ src/ipc/handlers/createFromTemplate.ts
@@ -5,10 +5,7 @@ import http from "isomorphic-git/http/node";
 import { app } from "electron";
 import { copyDirectoryRecursive } from "../utils/file_utils";
 import { readSettings } from "@/main/settings";
-import {
-  DEFAULT_TEMPLATE_ID,
-  getTemplateOrThrowAsync,
-} from "../utils/template_utils";
+import { getTemplateOrThrowAsync } from "../utils/template_utils";
 import log from "electron-log";
 
 const logger = log.scope("createFromTemplate");
@@ -19,7 +16,7 @@ export async function createFromTemplate({
   fullAppPath: string;
 }) {
   const settings = readSettings();
-  const templateId = settings.selectedTemplateId ?? DEFAULT_TEMPLATE_ID;
+  const templateId = settings.selectedTemplateId;
 
   if (templateId === "react") {
     await copyDirectoryRecursive(
