+++ src/main/settings.ts
@@ -5,7 +5,7 @@ import { UserSettingsSchema, type UserSettings, Secret } from "../lib/schemas";
 import { safeStorage } from "electron";
 import { v4 as uuidv4 } from "uuid";
 import log from "electron-log";
-import { DEFAULT_TEMPLATE_ID } from "@/ipc/utils/template_utils";
+import { DEFAULT_TEMPLATE_ID } from "@/shared/templates";
 
 const logger = log.scope("settings");
