+++ src/shared/templates.ts
@@ -15,3 +15,25 @@ export interface ApiTemplate {
   description: string;
   imageUrl: string;
 }
+
+export const DEFAULT_TEMPLATE_ID = "react";
+
+export const localTemplatesData: Template[] = [
+  {
+    id: "react",
+    title: "React.js Template",
+    description: "Uses React.js, Vite, Shadcn, Tailwind and TypeScript.",
+    imageUrl:
+      "https://github.com/user-attachments/assets/5b700eab-b28c-498e-96de-8649b14c16d9",
+    isOfficial: true,
+  },
+  {
+    id: "next",
+    title: "Next.js Template",
+    description: "Uses Next.js, React.js, Shadcn, Tailwind and TypeScript.",
+    imageUrl:
+      "https://github.com/user-attachments/assets/96258e4f-abce-4910-a62a-a9dff77965f2",
+    githubUrl: "https://github.com/dyad-sh/nextjs-template",
+    isOfficial: true,
+  },
+];
